# Ensure input, labels, and format are all in UTF-8 when parsing