--- a/src/parse.h
+++ b/src/parse.h
@@ -1418,10 +1418,11 @@ from_stream(std::basic_istream<CharT, Traits>& is,
 static
 inline
 void
-fill_formats(const cpp11::strings& src, std::vector<const char*>& dest) {
+fill_formats(const cpp11::strings& src, std::vector<std::string>& dest) {
     const r_ssize size = src.size();
     for (r_ssize i = 0; i < size; ++i) {
-        dest[i] = CHAR(src[i]);
+        std::string elt = src[i];
+        dest[i] = elt;
     }
 }