--- a/src/zoned-time.cpp
+++ b/src/zoned-time.cpp
@@ -429,7 +429,7 @@ static
 inline
 void
 zoned_parse_complete_one(std::istringstream& stream,
-                         const std::vector<const char*>& fmts,
+                         const std::vector<std::string>& fmts,
                          const std::pair<const std::string*, const std::string*>& month_names_pair,
                          const std::pair<const std::string*, const std::string*>& weekday_names_pair,
                          const std::pair<const std::string*, const std::string*>& ampm_names_pair,
@@ -448,7 +448,7 @@ zoned_parse_complete_one(std::istringstream& stream,
     stream.clear();
     stream.seekg(0);
 
-    const char* fmt = fmts[j];
+    const char* fmt = fmts[j].c_str();
 
     date::local_time<Duration> lt;
     std::string new_zone;
@@ -526,7 +526,7 @@ zoned_parse_complete_impl(const cpp11::strings& x,
   const r_ssize size = x.size();
   ClockDuration fields(size);
 
-  std::vector<const char*> fmts(format.size());
+  std::vector<std::string> fmts(format.size());
   rclock::fill_formats(format, fmts);
 
   char dmark;
@@ -639,7 +639,7 @@ static
 inline
 void
 zoned_parse_abbrev_one(std::istringstream& stream,
-                       const std::vector<const char*>& fmts,
+                       const std::vector<std::string>& fmts,
                        const std::pair<const std::string*, const std::string*>& month_names_pair,
                        const std::pair<const std::string*, const std::string*>& weekday_names_pair,
                        const std::pair<const std::string*, const std::string*>& ampm_names_pair,
@@ -656,7 +656,7 @@ zoned_parse_abbrev_one(std::istringstream& stream,
     stream.clear();
     stream.seekg(0);
 
-    const char* fmt = fmts[j];
+    const char* fmt = fmts[j].c_str();
 
     date::local_time<Duration> lt;
     std::string parsed_abbrev;
@@ -737,7 +737,7 @@ zoned_parse_abbrev_impl(const cpp11::strings& x,
   const r_ssize size = x.size();
   ClockDuration fields(size);
 
-  std::vector<const char*> fmts(format.size());
+  std::vector<std::string> fmts(format.size());
   rclock::fill_formats(format, fmts);
 
   char dmark;