Negligible performance impact, which is good

``` r
library(clock)


x <- rep("2019-01-01", 1e6)
bench::mark(year_month_day_parse(x), iterations = 10)

# before
#> Warning: Some expressions had a GC in every iteration; so filtering is disabled.
#> # A tibble: 1 x 6
#>   expression                   min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>              <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 year_month_day_parse(x)    387ms    396ms      2.45      12MB     3.92

# after
#> Warning: Some expressions had a GC in every iteration; so filtering is disabled.
#> # A tibble: 1 x 6
#>   expression                   min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>              <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 year_month_day_parse(x)    386ms    391ms      2.46      12MB     3.94


x <- rep("2019-01-01 01:02:03", 1e6)
bench::mark(naive_parse(x), iterations = 10)

# before
#> Warning: Some expressions had a GC in every iteration; so filtering is disabled.
#> # A tibble: 1 x 6
#>   expression          min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>     <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 naive_parse(x)    580ms    594ms      1.68    7.71MB     3.02

# after
#> Warning: Some expressions had a GC in every iteration; so filtering is disabled.
#> # A tibble: 1 x 6
#>   expression          min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>     <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 naive_parse(x)    579ms    592ms      1.68    7.71MB     3.03
```

<sup>Created on 2021-02-24 by the [reprex package](https://reprex.tidyverse.org) (v1.0.0)</sup>